% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pheno_vals.R
\name{calc_pheno_vals}
\alias{calc_pheno_vals}
\title{Cacluate STR mutator phenotypes}
\usage{
calc_pheno_vals(
  denovo_ri_gts,
  pheno = c("denovo_abundance", "denovo_perc_abundance", "proportion_expanded",
    "expand_delta_ru", "contract_delta_ru", "frac_D_block_denovo", "denovo_ri_len",
    "denovo_fou_len", "denovo_ri_len_expan", "denovo_fou_len_expan",
    "denovo_ri_len_contr", "denovo_fou_len_contr"),
  min_pts_per_phe = 1,
  max_denovo_strains_per_loc = 1000,
  ml = c("all", 2, 3, 4, 5, 6),
  gtloc_per_strain = NULL,
  verbose = FALSE
)
}
\arguments{
\item{denovo_ri_gts}{TODO}

\item{pheno}{TODO}

\item{min_pts_per_phe}{TODO}

\item{max_denovo_strains_per_loc}{TODO}

\item{ml}{TODO: motif length}

\item{gtloc_per_strain}{TODO}
}
\value{
tibble
}
\description{
Function for calculating various STR mutator phenotypes which could be used for QTL mapping
}
\examples{
pheno_vals = calc_pheno_vals(denovo_ri_gts, 'proportion_expanded', 1, 1000, ml = motif_len)
}
