% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_schemas}
\alias{db_schemas}
\title{Database schema}
\format{
a list of lists. Top level item are schema names. Next level are table names and finally field names.
\describe{
}
}
\source{

}
\usage{
db_schemas
}
\description{
Names of tables and fields within table for relevant sqlite databases
}
\keyword{datasets}
