% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denovo_strs}
\alias{denovo_strs}
\title{Denovo STRs}
\format{
a data frame with:
\describe{
\item{chr}{chromosome name}
\item{pos}{start position}
\item{end}{end position}
\item{RN_A}{A allele repeat count}
\item{RN_B}{B allele repeat count}
\item{RN_T}{Total repeat count}
\item{strain}{Strain name}
\item{founder}{Founder label B/D}
\item{founder_rn}{Repeat count in founder (single allele)}
\item{delta_fou}{Absolute difference in repeat count b/w RIL and founder (RU)}
\item{expand_sign}{Direction of difference +1/-1}
\item{expand_type}{Expansion/contraction}
}
}
\source{

}
\usage{
denovo_strs
}
\description{
RIL genotypes for denovo STRs. Homozygous denovos only
}
\keyword{datasets}
