% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dset_strains}
\alias{dset_strains}
\title{List of BXD strains featured in each expression dataset}
\format{
a data frame with:
\describe{
\item{GN}{Dataset id; e.g. GN105}
\item{strain}{BXD strain name; e.g. BXD001}
}
}
\source{

}
\usage{
dset_strains
}
\description{
Each dataset has its own unique set of strains available
}
\keyword{datasets}
