% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fasta_files}
\alias{fasta_files}
\title{fasta files}
\format{
a list of file paths:
\describe{
}
}
\source{

}
\usage{
fasta_files
}
\description{
Addresses of relevant fasta files
}
\keyword{datasets}
