% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{founder_blocks}
\alias{founder_blocks}
\title{Founder blocks from snps}
\format{
a data frame with:
\describe{
\item{strain}{Strain id; e.g. BXD001}
\item{chr}{Chromosome name}
\item{pos}{Start position bp}
\item{end}{End position bp}
\item{lab}{Founder label}
\item{n_loci}{Number of loci}
}
}
\source{

}
\usage{
founder_blocks
}
\description{
bedgraph stype dataframe with starts and stops of B/D inheritance blocks within
BXD
}
\keyword{datasets}
