% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gn_info}
\alias{gn_info}
\title{GeneNetwork dataset info}
\format{
a data frame with:
\describe{
\item{dset}{Dataset id; e.g. GN206}
\item{tissue}{Tissue type; e.g. eye_retina}
\item{tissue_group}{Tissue group; e.g. eye}
\item{tissue_id}{Tissue id; unique identifier if multiple datasets with same tissue; e.g. brain_hippocampus_1}
}
}
\source{

}
\usage{
gn_info
}
\description{
Tissue types for each GeneNetwork expression dataset
}
\keyword{datasets}
