% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtloc_per_strain}
\alias{gtloc_per_strain}
\title{Genotyped loci per BXD strain}
\format{
a data frame with:
\describe{
\item{strain}{Strain id; e.g. BXD001}
\item{n_loci}{Number of loci}
}
}
\source{

}
\usage{
gtloc_per_strain
}
\description{
Number of successfully genotyped STR loci per BXD strain
}
\keyword{datasets}
