% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_info}
\alias{motif_info}
\title{STR motif information}
\format{
a data frame with:
\describe{
\item{chr}{chromosome name}
\item{pos}{start position bp}
\item{end}{end position bp}
\item{motif_len}{length of the motif bp}
\item{motif}{motif identity}
\item{A}{number of As}
\item{C}{number of Cs}
\item{G}{number of Gs}
\item{T}{number of Ts}
\item{canon_motif}{canonical motif}
}
}
\source{

}
\usage{
motif_info
}
\description{
Properties of STR motifs
}
\keyword{datasets}
