% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qtl_data}
\alias{qtl_data}
\title{R/qtl2 formatted data}
\format{
a list of qtl2 objects:
\describe{
}
}
\source{

}
\usage{
qtl_data
}
\description{
genotype probabilities, physical map and strain kinship using SNPs
}
\keyword{datasets}
