% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_vcf}
\alias{query_vcf}
\title{Query vcf}
\usage{
query_vcf(
  vcf,
  loc_type = c("str", "snp")[1],
  query_type = c("wide", "long")[1],
  regions_file = NULL,
  samples = NULL,
  region = NULL,
  cstm_filt = NULL,
  verbose = FALSE
)
}
\arguments{
\item{vcf}{string}

\item{loc_type}{string}

\item{query_type}{string}

\item{regions_file}{string}

\item{samples}{string}

\item{region}{string}

\item{verbose}{logical}
}
\value{
tibble
}
\description{
Wrapper around bcftools to query data from vcf. Can read standard vcfs when
\code{loc_type} is snp or GangSTR vcfs when \code{loc_type} is str.
Can optionally specify a bed file with regions to only return variants at
at those coordinates or a region in the chr1:32038-34364 format.
If samples are specified, then data for only those samples can be returned.
query_type control whether awk is used to parse output of bcftools in long
format or if genotypes will be pulled into R in the wide format where
each sample is its own column
}
