% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{STR regions}
\format{
a data frame with:
\describe{
\item{chr}{chromosome name}
\item{pos}{start position bp}
\item{end}{end position bp}
}
}
\source{

}
\usage{
regions
}
\description{
a names list of bed files with STR coordinates
}
\keyword{datasets}
