% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gemma.R
\name{run_gemma}
\alias{run_gemma}
\title{Run GEMMA}
\usage{
run_gemma(
  geno,
  pheno,
  covar,
  grm,
  strains,
  covar_cols,
  gemma_exec,
  gemma_options,
  tmp_dir,
  tmp_file,
  permute = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{geno}{locus    al_A    al_B    BXD001    BXD002    BXD003
locA    C    T        12    5    14
locB    G    A        1    8    11}

\item{pheno}{strain    pheno1 pheno2
BXD001    1.3    12.2
BXD003    0.2    13.1
...
BXD222    2    20}

\item{covar}{strain    intercept    covar1    covar2
BXD001    1    1
BXD003    1    2
...
BXD222    1    2}

\item{grm}{matrix
strain BXD001 BXD002 ... BXD222
BXD001 1      0.3     0.89
BXD002
BXD222             1}

\item{strains}{vector of strains corresponding to those in geno, pheno and covar to be used.
Needs to be provided even without subselection to reorder the matrices.}

\item{gemma_exec}{string pointing to gemma executable}

\item{permute}{logical}

\item{verbose}{logical}
}
\value{
tibble
}
\description{
Wrapper around GEMMA from R
}
