% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sqlite_files}
\alias{sqlite_files}
\title{sqlite databases}
\format{
a list of file paths:
\describe{
}
}
\source{

}
\usage{
sqlite_files
}
\description{
Addresses of relevant sqlite databases
}
\keyword{datasets}
