% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strain_info}
\alias{strain_info}
\title{BXD strain info}
\format{
a data frame with:
\describe{
\item{long_name}{Long name for the strain; e.g. 4512-JFI-0361_BXD001_TyJ}
\item{short_name}{Short name for the strain; e.g. BXD001_TyJ_0361}
\item{bxd_id}{BXD id for the strain; e.g. BXD001}
\item{off_epoch}{Name of official epoch}
\item{gen_inbreeding}{Number of generations of inbreeding}
\item{gen_breeding}{Number of generations of breeding}
\item{cross_type}{Cross type}
}
}
\source{

}
\usage{
strain_info
}
\description{
Strain information about BXD strains
}
\keyword{datasets}
